requirements.txt
openai-agents
mcp
python-dotenv
